renamed some stuff, did some additional code moving around, removed unused dependencies
