+++ src/lib.rs
@@ -13,7 +13,7 @@ pub mod models {
 
 pub mod interface {
     pub mod argument_parser;
-    pub mod ui;
+    pub mod tui;
     pub mod state;
     pub mod log_options;
     pub mod log;
